Moving the DevTest validation into the azure package
