+++ azurerm/data_source_dev_test_lab.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
 
@@ -15,7 +16,7 @@ func dataSourceArmDevTestLab() *schema.Resource {
 			"name": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: validateDevTestLabName(),
+				ValidateFunc: azure.ValidateDevTestLabName(),
 			},
 
 			"location": locationForDataSourceSchema(),
