+++ azurerm/resource_arm_dev_test_lab.go
@@ -3,11 +3,11 @@ package azurerm
 import (
 	"fmt"
 	"log"
-	"regexp"
 
 	"github.com/Azure/azure-sdk-for-go/services/devtestlabs/mgmt/2016-05-15/dtl"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/helper/validation"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
 
@@ -26,7 +26,7 @@ func resourceArmDevTestLab() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: validateDevTestLabName(),
+				ValidateFunc: azure.ValidateDevTestLabName(),
 			},
 
 			"location": locationSchema(),
@@ -202,10 +202,4 @@ func resourceArmDevTestLabDelete(d *schema.ResourceData, meta interface{}) error
 	}
 
 	return err
-}
-
-func validateDevTestLabName() schema.SchemaValidateFunc {
-	return validation.StringMatch(
-		regexp.MustCompile("^[A-Za-z0-9_-]+$"),
-		"Lab Name can only include alphanumeric characters, underscores, hyphens.")
 }
