+++ azurerm/resource_arm_dev_test_lab_test.go
@@ -10,34 +10,6 @@ import (
 	"github.com/hashicorp/terraform/terraform"
 )
 
-func TestValidateDevTestLabName(t *testing.T) {
-	validNames := []string{
-		"valid-name",
-		"valid02-name",
-		"validName1",
-		"-validname1",
-		"valid_name",
-		"double-hyphen--valid",
-	}
-	for _, v := range validNames {
-		_, errors := validateDevTestLabName()(v, "example")
-		if len(errors) != 0 {
-			t.Fatalf("%q should be a valid Dev Test Lab Name: %q", v, errors)
-		}
-	}
-
-	invalidNames := []string{
-		"invalid!",
-		"!@£",
-	}
-	for _, v := range invalidNames {
-		_, errors := validateDevTestLabName()(v, "name")
-		if len(errors) == 0 {
-			t.Fatalf("%q should be an invalid Dev Test Lab Name", v)
-		}
-	}
-}
-
 func TestAccAzureRMDevTestLab_basic(t *testing.T) {
 	resourceName := "azurerm_dev_test_lab.test"
 	rInt := acctest.RandInt()
